Implement interface for generating proposals
